requirements.txt
requests
beautifulsoup4
pandas
PyYAML
